Reset form after submit comment
We now resetting the form after submitting a comment.
With the '#' we can create a lokale variable in the HTML Markup and then reference this in a class via the the @ViewChild('variableName') decorator. As this is a form tag we also imported the NgForm from @angular/forms and cast our variable as NgForm:

@ViewChild('commentForm') commentForm: NgForm;

so we have access to all the form methods.